+++ Agents/Home.Agents.Clara/Program.cs
@@ -13,27 +13,7 @@ namespace Home.Agents.Clara
 
         static void Main(string[] args)
         {
-            Console.WriteLine("Starting Clara");
-
-            string cn = Environment.GetEnvironmentVariable("APPCONFIG_CNSTRING");
-
-            if (cn != null)
-            {
-                Console.WriteLine("with config = " + cn);
-                ConfigurationSettingsHelper.Init(cn);
-            }
-            else
-                Console.WriteLine("without config");
-            Console.WriteLine("-----------------------");
-            var envs = Environment.GetEnvironmentVariables();
-            Console.WriteLine("Env vars : ");
-            foreach (var k in envs.Keys)
-            {
-                Console.Write(k.ToString().PadRight(35, ' '));
-                Console.Write(" : ");
-                Console.WriteLine(envs[k].ToString());
-            }
-            Console.WriteLine("-----------------------");
+            AgentHelper.WriteStartupMessage("Clara", typeof(Program).Assembly);
 
             AgentHelper.SetupLocaleFromServer("clara");
             AgentHelper.ReportStart("clara", "pim");
