+++ Agents/Home.Agents.Erina/Program.cs
@@ -10,26 +10,7 @@ namespace Home.Agents.Erina
 
         static void Main(string[] args)
         {
-            Console.WriteLine("Starting Erina");
-            string cn = Environment.GetEnvironmentVariable("APPCONFIG_CNSTRING");
-
-            if (cn != null)
-            {
-                Console.WriteLine("with config = " + cn);
-                ConfigurationSettingsHelper.Init(cn);
-            }
-            else
-                Console.WriteLine("without config");
-            Console.WriteLine("-----------------------");
-            var envs = Environment.GetEnvironmentVariables();
-            Console.WriteLine("Env vars : ");
-            foreach (var k in envs.Keys)
-            {
-                Console.Write(k.ToString().PadRight(35, ' '));
-                Console.Write(" : ");
-                Console.WriteLine(envs[k].ToString());
-            }
-            Console.WriteLine("-----------------------");
+            AgentHelper.WriteStartupMessage("Erina", typeof(Program).Assembly);
 
             AgentHelper.SetupLocaleFromServer("erina");
             AgentHelper.ReportStart("erina", "pim");
