+++ Agents/Home.Agents.Erza/Program.cs
@@ -13,27 +13,7 @@ namespace Home.Agents.Erza
 
         static void Main(string[] args)
         {
-            Console.WriteLine("Starting Erza");
-            string cn = Environment.GetEnvironmentVariable("APPCONFIG_CNSTRING");
-
-
-            if (cn != null)
-            {
-                Console.WriteLine("with config = " + cn);
-                ConfigurationSettingsHelper.Init(cn);
-            }
-            else
-                Console.WriteLine("without config");
-            Console.WriteLine("-----------------------");
-            var envs = Environment.GetEnvironmentVariables();
-            Console.WriteLine("Env vars : ");
-            foreach (var k in envs.Keys)
-            {
-                Console.Write(k.ToString().PadRight(35, ' '));
-                Console.Write(" : ");
-                Console.WriteLine(envs[k].ToString());
-            }
-            Console.WriteLine("-----------------------");
+            AgentHelper.WriteStartupMessage("Erza", typeof(Program).Assembly);
 
             DatabaseMaintenanceThread.CheckDb();
