+++ Agents/Home.Agents.Freeia/Program.cs
@@ -11,22 +11,11 @@ namespace Home.Agents.Freeia
 
         static void Main(string[] args)
         {
-            Console.WriteLine("Starting Freeia");
-            string cn = Environment.GetEnvironmentVariable("APPCONFIG_CNSTRING");
+            AgentHelper.WriteStartupMessage("Freeia", typeof(Program).Assembly);
 
             AgentHelper.SetupLocaleFromServer("freeia");
             AgentHelper.ReportStart("freeia", "monitoring");
 
-            Console.WriteLine("-----------------------");
-            var envs = Environment.GetEnvironmentVariables();
-            Console.WriteLine("Env vars : ");
-            foreach (var k in envs.Keys)
-            {
-                Console.Write(k.ToString().PadRight(35, ' '));
-                Console.Write(" : ");
-                Console.WriteLine(envs[k].ToString());
-            }
-            Console.WriteLine("-----------------------");
 
             MqttHandler.Start();
             FreeiaMessageHandler.Start();
