+++ Agents/Home.Agents.Gaia/Program.cs
@@ -8,46 +8,13 @@ namespace Home.Agents.Gaia
     class Program
     {
         public static bool _stop = false;
-    
+
         static void Main(string[] args)
         {
-            DeploymentHelper.DeployWebApp("security");
-
-            Console.WriteLine("Starting Gaïa");
-            string cn = Environment.GetEnvironmentVariable("APPCONFIG_CNSTRING");
-
-            if (cn != null)
-            {
-                Console.WriteLine("with config = " + cn);
-                ConfigurationSettingsHelper.Init(cn);
-            }
-            else
-                Console.WriteLine("without config");
-            Console.WriteLine("-----------------------");
-            var envs = Environment.GetEnvironmentVariables();
-            Console.WriteLine("Env vars : ");
-            foreach (var k in envs.Keys)
-            {
-                Console.Write(k.ToString().PadRight(35, ' '));
-                Console.Write(" : ");
-                Console.WriteLine(envs[k].ToString());
-            }
-            Console.WriteLine("-----------------------");
-
-            try
-            {
-                var config = KubernetesClientConfiguration.InClusterConfig();
-                Console.WriteLine($"host for k8s configuration : {config.Host}");
-                Console.WriteLine("-----------------------");
-
-            }
-            catch
-            {
-
-            }
+            AgentHelper.WriteStartupMessage("Gaïa", typeof(Program).Assembly);
 
 
-                AgentHelper.SetupLocaleFromServer("gaia");
+            AgentHelper.SetupLocaleFromServer("gaia");
             AgentHelper.ReportStart("gaia", "monitoring", "system");
